[VFS-483][RAM] Many suggestions to improve the RAM file provider. Use long for FS size limit.
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/vfs/trunk@1511561 13f79535-47bb-0310-9956-ffa450edef68

+++ core/src/main/java/org/apache/commons/vfs2/provider/ram/RamFileSystem.java
@@ -300,9 +300,9 @@ public class RamFileSystem extends AbstractFileSystem implements Serializable
     /**
      * @return Returns the size of the FileSystem
      */
-    int size()
+    long size()
     {
-        int size = 0;
+        long size = 0;
         synchronized (cache)
         {
             final Iterator<RamFileData> iter = cache.values().iterator();
